lots of saving tweaks, new parasite bases + more
save player placed ground items, save farm progress, new flesh walls, parasite base, raids and scouter waves get harder over time. New algorithm for random positions in radiuses. Lots of bug fixes.

+++ Assets/Scripts/Components/HealthManager.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 
 public class HealthManager : MonoBehaviour
 {
-    public int maxHealth;
-    public int currentHealth;
-    public int currentArmor = 0;
+    public float maxHealth;
+    public float currentHealth;
+    public float currentArmor = 0;
     public DamageArgs dmgArgs = new DamageArgs();
     public EventHandler<DamageArgs> OnDamageTaken;
     public EventHandler<DamageArgs> OnDeath;
